+ - Respect offsets in canned cycle moves.  #219 + - Fixed G53 warning. + - Fixed delayed offset update after M2 or M30 end of program. + - Handle multiple consecutive config resets correctly. + - Fixed log CPU usage problem introduced in v0.4.6. + - Show RPi temp on indicators page. + - Show red thermometer if RPi temp exceeds 80°C. + - Enforce 6A per motor channel peak current limit. + - Adjust config values above max or below min instead of resetting to default.

+++ src/py/bbctrl/Config.py
@@ -94,9 +94,7 @@ class Config(object):
         except:
             return False
 
-        if (('min' in template and value < template['min']) or
-            ('max' in template and template['max'] < value) or
-            ('values' in template and value not in template['values'])):
+        if 'values' in template and value not in template['values']:
             return False
 
         return True
@@ -107,6 +105,12 @@ class Config(object):
             if (not name in config or
                 not self._valid_value(template, config[name])):
                 config[name] = template['default']
+
+            elif 'max' in template and template['max'] < config[name]:
+                config[name] = template['max']
+
+            elif 'min' in template and config[name] < template['min']:
+                config[name] = template['min']
 
             if template['type'] == 'list':
                 config = config[name]
@@ -182,7 +186,7 @@ class Config(object):
 
 
     def reset(self):
-        os.unlink('config.json')
+        if os.path.exists('config.json'): os.unlink('config.json')
         self.reload()
         self.ctrl.preplanner.invalidate_all()
